First pass: Port to Python3, make code Flake8 clean
This is a _very_ rough cut at making this code work under Python 3. It
seems to work for my very simple use case, but there's a ton of stuff
in the codebase that I don't use and haven't tested. Also, much of the
documentation still references Python 2. I'm sure there are many other
problems.

Having said that, it is working for my purposes.

Of note: The old Python 2 version of TMDA was shipped with its own
versions of several of the standard Python "email" libraries. Those
have been deleted; we're using the standard libraries now.

My approach to porting it to Python 3 was to run Flake8 over the
entire code base and fix every warning or error that was reported, and
then just keep running the scripts repeatedly to uncover issues not
flagged by Flake8 and fix them as they occurred. I'm sure I've missed
many things.